semantic/requirements.txt
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.31.1
plotly==5.18.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-google-genai==0.0.5
openpyxl==3.1.2 